promocoes delete,update,getall adiciona tratamento de retorno

+++ curso1-intro-kotlin-springboot/src/main/kotlin/com/acme/tour/controller/PromocaoController.kt
@@ -2,6 +2,7 @@ package com.acme.tour.controller
 
 import com.acme.tour.model.Promocao
 import com.acme.tour.service.PromocaoService
+import org.apache.coyote.Response
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
@@ -16,9 +17,16 @@ class PromocaoController {
     lateinit var promocaoService: PromocaoService
 
     @GetMapping()
-    fun getAll(@RequestParam(required = false ,defaultValue = "") localFilter: String) =
-        this.promocaoService.searchByLocal(localFilter)
+    fun getAll(@RequestParam(required = false ,defaultValue = "") localFilter: String): ResponseEntity<List<Promocao>> {
+        var status = HttpStatus.OK
+        val listaPromocoes = this.promocaoService.searchByLocal(localFilter)
 
+        if (listaPromocoes.isEmpty()){
+            status = HttpStatus.NOT_FOUND
+        }
+
+        return ResponseEntity(listaPromocoes, status)
+    }
     @GetMapping("/{id}")
     fun getById(@PathVariable id: Long): ResponseEntity<Promocao?> {
         val promocao = this.promocaoService.getById(id)
@@ -35,12 +43,23 @@ class PromocaoController {
     }
 
     @DeleteMapping("/{id}")
-    fun delete(@PathVariable id: Long) {
-        this.promocaoService.delete(id)
+    fun delete(@PathVariable id: Long): ResponseEntity<Unit> {
+        var status = HttpStatus.NOT_FOUND
+        if (this.promocaoService.getById(id) != null) {
+            this.promocaoService.delete(id)
+            status = HttpStatus.ACCEPTED
+        }
+
+        return ResponseEntity(Unit, status)
     }
 
     @PutMapping("/{id}")
-    fun update(@PathVariable id: Long, @RequestBody promocao: Promocao) {
-        this.promocaoService.update(id, promocao)
+    fun update(@PathVariable id: Long, @RequestBody promocao: Promocao): ResponseEntity<Unit> {
+        var status = HttpStatus.NOT_FOUND
+        if (this.promocaoService.getById(id) != null){
+            this.promocaoService.update(id, promocao)
+            status = HttpStatus.ACCEPTED
+        }
+        return ResponseEntity(Unit,status)
     }
 }
